requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
google-generativeai>=0.8.5
bs4==0.0.2
python-dotenv==1.0.1
tiktoken==0.7.0
langchain_chroma>=0.1.4
langchainhub==0.1.20
sentence-transformers>=5.0.0
streamlit>=1.28.0